feat: add add attendees page
