+++ src/pages/main/Attendees/Attendees.tsx
@@ -1,22 +1,22 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useEffect, useState } from "react";
+import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
 
-import { deleteAttendee, useAttendees } from "../../api/attendee.ts";
-import AlertModal from "../../components/AlertModal.tsx";
-import AddAttendeesModal from "../../components/attendees/AddAttendeesModal.tsx";
-import AttendeesItem from "../../components/attendees/AttendeesItem.tsx";
-import { PrimaryButton, SecondaryButton } from "../../components/Button.tsx";
-import Checkbox from "../../components/Checkbox.tsx";
+import { deleteAttendee, useAttendees } from "../../../api/attendee.ts";
+import AlertModal from "../../../components/AlertModal.tsx";
+import AttendeesItem from "../../../components/attendees/AttendeesItem.tsx";
+import { PrimaryButton, SecondaryButton } from "../../../components/Button.tsx";
+import Checkbox from "../../../components/Checkbox.tsx";
 import {
   SessionTable,
   SessionTableHead,
   SessionTableHeadItem,
-} from "../../components/sessions/SessionTable.tsx";
-import TitleBar from "../../components/TitleBar.tsx";
-import useModalState from "../../hooks/modal.tsx";
-import { useClassId } from "../../hooks/urlParse.tsx";
-import theme from "../../styles/Theme.tsx";
+} from "../../../components/sessions/SessionTable.tsx";
+import TitleBar from "../../../components/TitleBar.tsx";
+import useModalState from "../../../hooks/modal.tsx";
+import { useClassId } from "../../../hooks/urlParse.tsx";
+import theme from "../../../styles/Theme.tsx";
 
 interface CheckedState {
   [key: number]: boolean;
@@ -73,7 +73,11 @@ function Attendees() {
   }, [checkedState, attendees]);
 
   // 유저 추가 관련
-  const [addModalState, openAddModal, closeAddModal] = useModalState();
+  const navigate = useNavigate();
+
+  const handleAddAttendees = () => {
+    navigate(`/class/${classId}/attendees/add`);
+  };
 
   // 유저 삭제 관련
   const [deleteModalState, openDeleteModal, closeDeleteModal] = useModalState();
@@ -96,7 +100,7 @@ function Attendees() {
     <>
       <Container>
         <TitleBar label="Attendees">
-          <PrimaryButton onClick={openAddModal}>
+          <PrimaryButton onClick={handleAddAttendees}>
             Add New Attendees
           </PrimaryButton>
         </TitleBar>
@@ -163,7 +167,6 @@ function Attendees() {
         </ContentContainer>
       </Container>
       {/* 모달 */}
-      <AddAttendeesModal state={addModalState} onCancel={closeAddModal} />
       <AlertModal
         state={deleteModalState}
         type="delete"
